@@ -69,7 +69,7 @@ RSpec.describe YAVDB::Sources::SnykIO::Client do
     it 'should have correct properties for SNYK-JAVA-ORGJBPM-31602' do
       vuln = advisories.select { |v| v.id == 'snykio:maven:org.jbpm:jbpm-designer-client:31602' }.first
       expect(vuln).to have_attributes(:title => 'Cross-site Scripting (XSS)')
-      expect(vuln).to have_attributes(:description => "Affected versions of [`org.jbpm:jbpm-designer-client`][1] are vulnerable\nto Cross-site Scripting (XSS).\n\n\n\n[1]: https://jbpm.org\n\\nCross-Site Scripting (XSS) attacks occur when an attacker tricks a\nuser’s browser to execute malicious JavaScript code in the context of a\nvictim’s domain. Such scripts can steal the user’s session cookies for\nthe domain, scrape or modify its content, and perform or modify actions\non the user’s behalf, actions typically blocked by the browser’s Same\nOrigin Policy.\n\nThese attacks are possible by escaping the context of the web\napplication and injecting malicious scripts in an otherwise trusted\nwebsite. These scripts can introduce additional attributes (say, a\n\\\"new\\\" option in a dropdown list or a new link to a malicious site) and\ncan potentially execute code on the clients side, unbeknown to the\nvictim. This occurs when characters like `<` `>` `\"` `'` are not escaped\nproperly.\n\nThere are a few types of XSS:\n\n* **Persistent XSS** is an attack in which the malicious code persists\n  into the web app’s database.\n* **Reflected XSS** is an which the website echoes back a portion of the\n  request. The attacker needs to trick the user into clicking a\n  malicious link (for instance through a phishing email or malicious JS\n  on another page), which triggers the XSS attack.\n* **DOM-based XSS** is an that occurs purely in the browser when\n  client-side JavaScript echoes back a portion of the URL onto the page.\n  DOM-Based XSS is notoriously hard to detect, as the server never gets\n  a chance to see the attack taking place.\n\n")
+      expect(vuln).to have_attributes(:description => "Affected versions of [`org.jbpm:jbpm-designer-client`][1] are vulnerable\nto Cross-site Scripting (XSS).\n\n\n\n[1]: https://jbpm.org\n\\nA cross-site scripting attack occurs when the attacker tricks a\nlegitimate web-based application or site to accept a request as\noriginating from a trusted source.\n\nThis is done by escaping the context of the web application; the web\napplication then delivers that data to its users along with other\ntrusted dynamic content, without validating it. The browser unknowingly\nexecutes malicious script on the client side (through client-side\nlanguages; usually JavaScript or HTML) in order to perform actions that\nare otherwise typically blocked by the browser’s Same Origin Policy.\n\nֿInjecting malicious code is the most prevalent manner by which XSS is\nexploited; for this reason, escaping characters in order to prevent this\nmanipulation is the top method for securing code against this\nvulnerability.\n\nEscaping means that the application is coded to mark key characters, and\nparticularly key characters included in user input, to prevent those\ncharacters from being interpreted in a dangerous context. For example,\nin HTML, `<` can be coded as `&lt`; and `>` can be coded as `&gt`; in\norder to be interpreted and displayed as themselves in text, while\nwithin the code itself, they are used for HTML tags. If malicious\ncontent is injected into an application that escapes special characters\nand that malicious content uses `<` and `>` as HTML tags, those\ncharacters are nonetheless not interpreted as HTML tags by the browser\nif they’ve been correctly escaped in the application code and in this\nway the attempted attack is diverted.\n\nThe most prominent use of XSS is to steal cookies (source: OWASP\nHttpOnly) and hijack user sessions, but XSS exploits have been used to\nexpose sensitive information, enable access to privileged services and\nfunctionality and deliver malware.\n\n### Types of attacks   {#types-of-attacks}\n\nThere are a few methods by which XSS can be manipulated:\n\n| Type | Origin | Description |\n|----------\n| **Stored** | Server | The malicious code is inserted in the application (usually as a link) by the attacker. The code is activated every time a user clicks the link. |\n| **Reflected** | Server | The attacker delivers a malicious link externally from the vulnerable web site application to a user. When clicked, malicious code is sent to the vulnerable web site, which reflects the attack back to the user’s browser. |\n| **DOM-based** | Client | The attacker forces the user’s browser to render a malicious page. The data in the page itself delivers the cross-site scripting data. |\n| **Mutated** |  | The attacker injects code that appears safe, but is then rewritten and modified by the browser, while parsing the markup. An example is rebalancing unclosed quotation marks or even adding quotation marks to unquoted parameters. |\n\n### Affected environments   {#affected-environments}\n\nThe following environments are susceptible to an XSS attack:\n\n* Web servers\n* Application servers\n* Web application environments\n\n### How to prevent   {#how-to-prevent}\n\nThis section describes the top best practices designed to specifically\nprotect your code:\n\n* Sanitize data input in an HTTP request before reflecting it back,\n  ensuring all data is validated, filtered or escaped before echoing\n  anything back to the user, such as the values of query parameters\n  during searches.\n* Convert special characters such as `?`, `&`, `/`, `<`, `>` and spaces\n  to their respective HTML or URL encoded equivalents.\n* Give users the option to disable client-side scripts.\n* Redirect invalid requests.\n* Detect simultaneous logins, including those from two separate IP\n  addresses, and invalidate those sessions.\n* Use and enforce a Content Security Policy (source: Wikipedia) to\n  disable any features that might be manipulated for an XSS attack.\n* Read the documentation for any of the libraries referenced in your\n  code to understand which elements allow for embedded HTML.\n\n")
       expect(vuln).to have_attributes(:affected_package => 'org.jbpm:jbpm-designer-client')
       expect(vuln).to have_attributes(:vulnerable_versions => ['[,6.3.3)'])
       expect(vuln).to have_attributes(:severity => 'medium')
@@ -87,7 +87,7 @@ RSpec.describe YAVDB::Sources::SnykIO::Client do
     it 'should have correct properties for npm:xlsx:20180222' do
       vuln = advisories.select { |v| v.id == 'snykio:npm:xlsx:20180222' }.first
       expect(vuln).to have_attributes(:title => 'Regular Expression Denial of Service (ReDoS)')
-      expect(vuln).to have_attributes(:description => "[`xlsx`][1] is a parser and writer for various spreadsheet formats.\n\nAffected versions of this package are vulnerable to Regular Expression\nDenial of Service (ReDoS) attacks. This can cause an impact of about 10\nseconds matching time for data 40k characters long.\n\n\n\n[1]: https://www.npmjs.com/package/xlsx\n\\nDenial of Service (DoS) describes a family of attacks, all aimed at\nmaking a system inaccessible to its original and legitimate users. There\nare many types of DoS attacks, ranging from trying to clog the network\npipes to the system by generating a large volume of traffic from many\nmachines (a Distributed Denial of Service - DDoS - attack) to sending\ncrafted requests that cause a system to crash or take a disproportional\namount of time to process.\n\nThe Regular expression Denial of Service (ReDoS) is a type of Denial of\nService attack. Regular expressions are incredibly powerful, but they\naren\\'t very intuitive and can ultimately end up making it easy for\nattackers to take your site down.\n\nLet’s take the following regular expression as an example:\n\n    regex = /A(B|C+)+D/\n\nThis regular expression accomplishes the following:\n\n* `A` The string must start with the letter \\'A\\'\n* `(B|C+)+` The string must then follow the letter A with either the\n  letter \\'B\\' or some number of occurrences of the letter \\'C\\' (the\n  `+` matches one or more times). The `+` at the end of this section\n  states that we can look for one or more matches of this section.\n* `D` Finally, we ensure this section of the string ends with a \\'D\\'\n\nThe expression would match inputs such as `ABBD`, `ABCCCCD`, `ABCBCCCD`\nand `ACCCCCD`\n\nIt most cases, it doesn\\'t take very long for a regex engine to find a\nmatch:\n\n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCD\")'\n    0.04s user 0.01s system 95% cpu 0.052 total\n    \n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCX\")'\n    1.79s user 0.02s system 99% cpu 1.812 total\n\nThe entire process of testing it against a 30 characters long string\ntakes around ~52ms. But when given an invalid string, it takes nearly\ntwo seconds to complete the test, over ten times as long as it took to\ntest a valid string. The dramatic difference is due to the way regular\nexpressions get evaluated.\n\nMost Regex engines will work very similarly (with minor differences).\nThe engine will match the first possible way to accept the current\ncharacter and proceed to the next one. If it then fails to match the\nnext one, it will backtrack and see if there was another way to digest\nthe previous character. If it goes too far down the rabbit hole only to\nfind out the string doesn’t match in the end, and if many characters\nhave multiple valid regex paths, the number of backtracking steps can\nbecome very large, resulting in what is known as *catastrophic\nbacktracking*.\n\nLet\\'s look at how our expression runs into this problem, using a\nshorter string: \\\"ACCCX\\\". While it seems fairly straightforward, there\nare still four different ways that the engine could match those three\nC\\'s:\n\n1.  CCC\n2.  CC+C\n3.  C+CC\n4.  C+C+C.\n\nThe engine has to try each of those combinations to see if any of them\npotentially match against the expression. When you combine that with the\nother steps the engine must take, we can use [RegEx 101 debugger][1] to\nsee the engine has to take a total of 38 steps before it can determine\nthe string doesn\\'t match.\n\nFrom there, the number of steps the engine must use to validate a string\njust continues to grow.\n\n| String | Number of C\\'s | Number of steps |\n|----------\n| ACCCX | 3 | 38 |\n| ACCCCX | 4 | 71 |\n| ACCCCCX | 5 | 136 |\n| ACCCCCCCCCCCCCCX | 14 | 65,553 |\n\nBy the time the string includes 14 C\\'s, the engine has to take over\n65,000 steps just to see if the string is valid. These extreme\nsituations can cause them to work very slowly (exponentially related to\ninput size, as shown above), allowing an attacker to exploit this and\ncan cause the service to excessively consume CPU, resulting in a Denial\nof Service.\n\n\n\n[1]: https://regex101.com/debugger\n")
+      expect(vuln).to have_attributes(:description => "[`xlsx`][1] is a parser and writer for various spreadsheet formats.\n\nAffected versions of this package are vulnerable to Regular Expression\nDenial of Service (ReDoS) attacks. This can cause an impact of about 10\nseconds matching time for data 40k characters long.\n\n\n\n[1]: https://www.npmjs.com/package/xlsx\n\\nDenial of Service (DoS) describes a family of attacks, all aimed at\nmaking a system inaccessible to its original and legitimate users. There\nare many types of DoS attacks, ranging from trying to clog the network\npipes to the system by generating a large volume of traffic from many\nmachines (a Distributed Denial of Service - DDoS - attack) to sending\ncrafted requests that cause a system to crash or take a disproportional\namount of time to process.\n\nThe Regular expression Denial of Service (ReDoS) is a type of Denial of\nService attack. Regular expressions are incredibly powerful, but they\naren\\'t very intuitive and can ultimately end up making it easy for\nattackers to take your site down.\n\nLet’s take the following regular expression as an example:\n\n    regex = /A(B|C+)+D/\n{: .language-js}\n\nThis regular expression accomplishes the following:\n\n* `A` The string must start with the letter \\'A\\'\n* `(B|C+)+` The string must then follow the letter A with either the\n  letter \\'B\\' or some number of occurrences of the letter \\'C\\' (the\n  `+` matches one or more times). The `+` at the end of this section\n  states that we can look for one or more matches of this section.\n* `D` Finally, we ensure this section of the string ends with a \\'D\\'\n\nThe expression would match inputs such as `ABBD`, `ABCCCCD`, `ABCBCCCD`\nand `ACCCCCD`\n\nIt most cases, it doesn\\'t take very long for a regex engine to find a\nmatch:\n\n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCD\")'\n    0.04s user 0.01s system 95% cpu 0.052 total\n    \n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCX\")'\n    1.79s user 0.02s system 99% cpu 1.812 total\n{: .language-bash}\n\nThe entire process of testing it against a 30 characters long string\ntakes around ~52ms. But when given an invalid string, it takes nearly\ntwo seconds to complete the test, over ten times as long as it took to\ntest a valid string. The dramatic difference is due to the way regular\nexpressions get evaluated.\n\nMost Regex engines will work very similarly (with minor differences).\nThe engine will match the first possible way to accept the current\ncharacter and proceed to the next one. If it then fails to match the\nnext one, it will backtrack and see if there was another way to digest\nthe previous character. If it goes too far down the rabbit hole only to\nfind out the string doesn’t match in the end, and if many characters\nhave multiple valid regex paths, the number of backtracking steps can\nbecome very large, resulting in what is known as *catastrophic\nbacktracking*.\n\nLet\\'s look at how our expression runs into this problem, using a\nshorter string: \\\"ACCCX\\\". While it seems fairly straightforward, there\nare still four different ways that the engine could match those three\nC\\'s:\n\n1.  CCC\n2.  CC+C\n3.  C+CC\n4.  C+C+C.\n\nThe engine has to try each of those combinations to see if any of them\npotentially match against the expression. When you combine that with the\nother steps the engine must take, we can use [RegEx 101 debugger][1] to\nsee the engine has to take a total of 38 steps before it can determine\nthe string doesn\\'t match.\n\nFrom there, the number of steps the engine must use to validate a string\njust continues to grow.\n\n| String | Number of C\\'s | Number of steps |\n|----------\n| ACCCX | 3 | 38 |\n| ACCCCX | 4 | 71 |\n| ACCCCCX | 5 | 136 |\n| ACCCCCCCCCCCCCCX | 14 | 65,553 |\n\nBy the time the string includes 14 C\\'s, the engine has to take over\n65,000 steps just to see if the string is valid. These extreme\nsituations can cause them to work very slowly (exponentially related to\ninput size, as shown above), allowing an attacker to exploit this and\ncan cause the service to excessively consume CPU, resulting in a Denial\nof Service.\n\n\n\n[1]: https://regex101.com/debugger\n")
       expect(vuln).to have_attributes(:affected_package => 'xlsx')
       expect(vuln).to have_attributes(:vulnerable_versions => ['<0.12.2'])
       expect(vuln).to have_attributes(:severity => 'low')
@@ -195,7 +195,7 @@ RSpec.describe YAVDB::Sources::SnykIO::Client do
     it 'should have correct properties for SNYK-DOTNET-MIME-60187' do
       vuln = advisories.select { |v| v.id == 'snykio:nuget:mime:60187' }.first
       expect(vuln).to have_attributes(:title => 'Regular Expression Denial of Service (ReDoS)')
-      expect(vuln).to have_attributes(:description => "[`mime`][1] is a comprehensive, compact MIME type module.\n\nAffected versions of this package are vulnerable to Regular expression\nDenial of Service (ReDoS). It uses regex the following regex\n`/.*[\\.\\/\\\\]/` in its lookup, which can cause a slowdown of 2 seconds\nfor 50k characters.\n\n\n\n[1]: https://www.npmjs.com/package/mime\n\\nDenial of Service (DoS) describes a family of attacks, all aimed at\nmaking a system inaccessible to its original and legitimate users. There\nare many types of DoS attacks, ranging from trying to clog the network\npipes to the system by generating a large volume of traffic from many\nmachines (a Distributed Denial of Service - DDoS - attack) to sending\ncrafted requests that cause a system to crash or take a disproportional\namount of time to process.\n\nThe Regular expression Denial of Service (ReDoS) is a type of Denial of\nService attack. Regular expressions are incredibly powerful, but they\naren\\'t very intuitive and can ultimately end up making it easy for\nattackers to take your site down.\n\nLet’s take the following regular expression as an example:\n\n    regex = /A(B|C+)+D/\n\nThis regular expression accomplishes the following:\n\n* `A` The string must start with the letter \\'A\\'\n* `(B|C+)+` The string must then follow the letter A with either the\n  letter \\'B\\' or some number of occurrences of the letter \\'C\\' (the\n  `+` matches one or more times). The `+` at the end of this section\n  states that we can look for one or more matches of this section.\n* `D` Finally, we ensure this section of the string ends with a \\'D\\'\n\nThe expression would match inputs such as `ABBD`, `ABCCCCD`, `ABCBCCCD`\nand `ACCCCCD`\n\nIt most cases, it doesn\\'t take very long for a regex engine to find a\nmatch:\n\n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCD\")'\n    0.04s user 0.01s system 95% cpu 0.052 total\n    \n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCX\")'\n    1.79s user 0.02s system 99% cpu 1.812 total\n\nThe entire process of testing it against a 30 characters long string\ntakes around ~52ms. But when given an invalid string, it takes nearly\ntwo seconds to complete the test, over ten times as long as it took to\ntest a valid string. The dramatic difference is due to the way regular\nexpressions get evaluated.\n\nMost Regex engines will work very similarly (with minor differences).\nThe engine will match the first possible way to accept the current\ncharacter and proceed to the next one. If it then fails to match the\nnext one, it will backtrack and see if there was another way to digest\nthe previous character. If it goes too far down the rabbit hole only to\nfind out the string doesn’t match in the end, and if many characters\nhave multiple valid regex paths, the number of backtracking steps can\nbecome very large, resulting in what is known as *catastrophic\nbacktracking*.\n\nLet\\'s look at how our expression runs into this problem, using a\nshorter string: \\\"ACCCX\\\". While it seems fairly straightforward, there\nare still four different ways that the engine could match those three\nC\\'s:\n\n1.  CCC\n2.  CC+C\n3.  C+CC\n4.  C+C+C.\n\nThe engine has to try each of those combinations to see if any of them\npotentially match against the expression. When you combine that with the\nother steps the engine must take, we can use [RegEx 101 debugger][1] to\nsee the engine has to take a total of 38 steps before it can determine\nthe string doesn\\'t match.\n\nFrom there, the number of steps the engine must use to validate a string\njust continues to grow.\n\n| String | Number of C\\'s | Number of steps |\n|----------\n| ACCCX | 3 | 38 |\n| ACCCCX | 4 | 71 |\n| ACCCCCX | 5 | 136 |\n| ACCCCCCCCCCCCCCX | 14 | 65,553 |\n\nBy the time the string includes 14 C\\'s, the engine has to take over\n65,000 steps just to see if the string is valid. These extreme\nsituations can cause them to work very slowly (exponentially related to\ninput size, as shown above), allowing an attacker to exploit this and\ncan cause the service to excessively consume CPU, resulting in a Denial\nof Service.\n\n\n\n[1]: https://regex101.com/debugger\n")
+      expect(vuln).to have_attributes(:description => "[`mime`][1] is a comprehensive, compact MIME type module.\n\nAffected versions of this package are vulnerable to Regular expression\nDenial of Service (ReDoS). It uses regex the following regex\n`/.*[\\.\\/\\\\]/` in its lookup, which can cause a slowdown of 2 seconds\nfor 50k characters.\n\n\n\n[1]: https://www.npmjs.com/package/mime\n\\nDenial of Service (DoS) describes a family of attacks, all aimed at\nmaking a system inaccessible to its original and legitimate users. There\nare many types of DoS attacks, ranging from trying to clog the network\npipes to the system by generating a large volume of traffic from many\nmachines (a Distributed Denial of Service - DDoS - attack) to sending\ncrafted requests that cause a system to crash or take a disproportional\namount of time to process.\n\nThe Regular expression Denial of Service (ReDoS) is a type of Denial of\nService attack. Regular expressions are incredibly powerful, but they\naren\\'t very intuitive and can ultimately end up making it easy for\nattackers to take your site down.\n\nLet’s take the following regular expression as an example:\n\n    regex = /A(B|C+)+D/\n{: .language-js}\n\nThis regular expression accomplishes the following:\n\n* `A` The string must start with the letter \\'A\\'\n* `(B|C+)+` The string must then follow the letter A with either the\n  letter \\'B\\' or some number of occurrences of the letter \\'C\\' (the\n  `+` matches one or more times). The `+` at the end of this section\n  states that we can look for one or more matches of this section.\n* `D` Finally, we ensure this section of the string ends with a \\'D\\'\n\nThe expression would match inputs such as `ABBD`, `ABCCCCD`, `ABCBCCCD`\nand `ACCCCCD`\n\nIt most cases, it doesn\\'t take very long for a regex engine to find a\nmatch:\n\n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCD\")'\n    0.04s user 0.01s system 95% cpu 0.052 total\n    \n    $ time node -e '/A(B|C+)+D/.test(\"ACCCCCCCCCCCCCCCCCCCCCCCCCCCCX\")'\n    1.79s user 0.02s system 99% cpu 1.812 total\n{: .language-bash}\n\nThe entire process of testing it against a 30 characters long string\ntakes around ~52ms. But when given an invalid string, it takes nearly\ntwo seconds to complete the test, over ten times as long as it took to\ntest a valid string. The dramatic difference is due to the way regular\nexpressions get evaluated.\n\nMost Regex engines will work very similarly (with minor differences).\nThe engine will match the first possible way to accept the current\ncharacter and proceed to the next one. If it then fails to match the\nnext one, it will backtrack and see if there was another way to digest\nthe previous character. If it goes too far down the rabbit hole only to\nfind out the string doesn’t match in the end, and if many characters\nhave multiple valid regex paths, the number of backtracking steps can\nbecome very large, resulting in what is known as *catastrophic\nbacktracking*.\n\nLet\\'s look at how our expression runs into this problem, using a\nshorter string: \\\"ACCCX\\\". While it seems fairly straightforward, there\nare still four different ways that the engine could match those three\nC\\'s:\n\n1.  CCC\n2.  CC+C\n3.  C+CC\n4.  C+C+C.\n\nThe engine has to try each of those combinations to see if any of them\npotentially match against the expression. When you combine that with the\nother steps the engine must take, we can use [RegEx 101 debugger][1] to\nsee the engine has to take a total of 38 steps before it can determine\nthe string doesn\\'t match.\n\nFrom there, the number of steps the engine must use to validate a string\njust continues to grow.\n\n| String | Number of C\\'s | Number of steps |\n|----------\n| ACCCX | 3 | 38 |\n| ACCCCX | 4 | 71 |\n| ACCCCCX | 5 | 136 |\n| ACCCCCCCCCCCCCCX | 14 | 65,553 |\n\nBy the time the string includes 14 C\\'s, the engine has to take over\n65,000 steps just to see if the string is valid. These extreme\nsituations can cause them to work very slowly (exponentially related to\ninput size, as shown above), allowing an attacker to exploit this and\ncan cause the service to excessively consume CPU, resulting in a Denial\nof Service.\n\n\n\n[1]: https://regex101.com/debugger\n")
       expect(vuln).to have_attributes(:affected_package => 'mime')
       expect(vuln).to have_attributes(:vulnerable_versions => ['[,1.4.1), [2,2.0.3)'])
       expect(vuln).to have_attributes(:severity => 'low')
