@@ -16,6 +16,6 @@
 
 module YAVDB
 
-  VERSION = '0.5.1'
+  VERSION = '0.5.2'
 
 end
